First Commit
First Commit

Game is a hide and seek game, based off the project in Head First C#
Page 332 designed to practice object oriented programming and
inheritence.  It has been modified some so that every room has a
description, and so that you're finding a variable number of children
with random names instead of just one.